# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Delivery API",
        "description": "API REST para sistema de gestión de pedidos y entregas",
        "contact": {
            "email": "admin@delivery.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Servidor de desarrollo"
        }
    ],
    "paths": {
...
        "/paquetes": {
            "get": {
                "tags": [
                    "Paquetes"
                ],
                "summary": "Listar paquetes",
                "description": "Obtiene una lista paginada de paquetes con opciones de filtrado",
                "operationId": "fcc589560a15ae1ac2b9033b36471ed5",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número de página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Elementos por página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    },
                    {
                        "name": "camioneros_id",
                        "in": "query",
                        "description": "Filtrar por ID de camionero",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "estados_paquetes_id",
                        "in": "query",
                        "description": "Filtrar por ID de estado",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de paquetes obtenida exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 50
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Paquetes"
                ],
                "summary": "Crear paquete",
                "description": "Crea un nuevo paquete con sus detalles. Requiere rol admin.",
                "operationId": "9abc0e946f172cd9b670c98531660810",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "camioneros_id",
                                    "estados_paquetes_id",
                                    "direccion"
                                ],
                                "properties": {
                                    "camioneros_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "estados_paquetes_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "direccion": {
                                        "type": "string",
                                        "maxLength": 25,
                                        "example": "Calle 123 #45-67"
                                    },
                                    "detalles": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "tipo_mercancia_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "dimencion": {
                                                    "type": "string",
                                                    "maxLength": 45,
                                                    "example": "50x30x40 cm"
                                                },
                                                "peso": {
                                                    "type": "string",
                                                    "maxLength": 45,
                                                    "example": "15 kg"
                                                },
                                                "fecha_entrega": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2025-12-01"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Paquete creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Errores de validación"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado - Se requiere rol admin"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/paquetes/{id}": {
            "get": {
                "tags": [
                    "Paquetes"
                ],
                "summary": "Obtener paquete",
                "description": "Obtiene los detalles de un paquete específico",
                "operationId": "425a12f3e1c1357217e49f31b3283b16",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del paquete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paquete obtenido exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Paquete no encontrado"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Paquetes"
                ],
                "summary": "Actualizar paquete",
                "description": "Actualiza un paquete existente. Requiere rol admin.",
                "operationId": "99b16ef37e4fe00f26347cf1d5602664",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del paquete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "camioneros_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "estados_paquetes_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "direccion": {
                                        "type": "string",
                                        "maxLength": 25,
                                        "example": "Nueva Calle 456"
                                    },
                                    "detalles": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "tipo_mercancia_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "dimencion": {
                                                    "type": "string",
                                                    "example": "60x40x50 cm"
                                                },
                                                "peso": {
                                                    "type": "string",
                                                    "example": "20 kg"
                                                },
                                                "fecha_entrega": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2025-12-05"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Paquete actualizado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Paquete no encontrado"
                    },
                    "422": {
                        "description": "Errores de validación"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado - Se requiere rol admin"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Paquetes"
                ],
                "summary": "Eliminar paquete",
                "description": "Elimina un paquete. Requiere rol admin.",
                "operationId": "8cd088e7bc25b4853bcc28f39d9f2c09",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del paquete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paquete eliminado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Eliminado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Paquete no encontrado"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado - Se requiere rol admin"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}